change from double to BigDecimal

+++ src/main/java/com/example/demo/CalculatorController.java
@@ -2,15 +2,16 @@ package com.example.demo;
 
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
 @RestController
 class CalculatorController {
-    List<Double> numbers = new ArrayList<>();
+    List<BigDecimal> numbers = new ArrayList<>();
 
     @PostMapping("calculator")
-    void postNumber(@RequestBody double number){
+    void postNumber(@RequestBody BigDecimal number){
         numbers.add(number);
     }
 
@@ -20,7 +21,7 @@ class CalculatorController {
     }
 
     @GetMapping("numbers")
-    List<Double> getNumbers(){
+    List<BigDecimal> getNumbers(){
         return numbers;
     }
 }
